requirements.txt
numpy
matplotlib
hikvisionapi
opencv-contrib-python
torch
torchvision
pygsheets
tdqm
psutil
scipy>=1.4.1
torch>=1.8
torchvision>=0.9
PyYAML>=5.4
seaborn
sklearn
pandas
requests
IPython
oauth2client
yolov5